iio: ad_adc: Add CN0363 support
Add support for the FPGA reference design for the CN0363.

The CN0363 uses the AD7173 as a analog front-end to do the
analog-to-digital conversion of the signal. The result is then fed into a
processing pipeline which outputs the result and multiple intermediate
results. Those results are copied to main memory via a DMA. This driver
enables/disables the front-end and the DMA when required as well as handles
the configuration of the data sequencer.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>