iio: axi_jesd204_rx: Automatically restart link on lane desynchronization
Setup a watchdog function that checks the state of each lane of the link
once a second. If a link has been established and payload data is being
transmitted but one of the lanes is detected to have entered a error state
restart the link (and hope the error has resolved itself).

Ideally we'd restart restart all components of the link to ensure that link
is reestablish cleanly, but we are not quite there yet.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>