iio: ad9467: Don't write 'register transfer' register for AD9680 and similar
According to the datasheet the AD9680 and related devices do not have a
'register transfer' register and all register writes are applied
immediately. Experimentation confirms that behavior is the same whether a
write to this register is done or not.

So remove all writes to this register for the AD9680 and related devices.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>