iio: ad9144: Add support for SPI 4-wire mode
The AD9144 supports both SPI 3-wire and 4-wire mode. The driver currently
configures the device for 3-wire mode by default. This is a bit unfortunate
since in the Linux SPI driver model 4-wire mode is the default and 3-wire
mode needs to be explicitly enabled.

Unfortunately the current behavior of the driver can not be changed without
breaking backwards compatibility. For this reason introduce the
non-standard 'adi,spi-4wire' property.

This is very much very wrong, but the only way to preserve backwards
compatibility. This non-standard property should never ever be copied to
new drivers.

And maybe one day when enough time has passed it is possible to update the
AD9144 driver to use standard default behavior.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>