net: macb: Avoid probe error if more than one phy is available
In the case where more than one phy is available on the same mdio bus, do
not return error on probe even if the MAC cannot connect to it's own
phydev.
This allows for another MAC to connect to one of the phydevs defined
in this bus. Take the dts example:

ethernet@broken {
	phy-handle = <&phy0>;

	phy0: phy@0 {
		device_type = "ethernet-phy";
		reg = <0>;
	};

	phy1: phy@1 {
		device_type = "ethernet-phy";
		reg = <1>;
	};
};

ethernet@other {
	phy-handle = <&phy1>;
};

With this patch, even if phy0 is broken (eg: faulty HW), phy1 will
still work...

The case described above, can be found on carrier-boards where there is one
PHY on the carrier and one on a board connected via FMC. The MDIO bus is
controlled by one of the MACs, and there are 2 PHYs connected to the same
MDIO bus.

This problem is also due to the fact that the MDIO control/logic is part of
the macb driver. A better alternative would be to split the MDIO logic into
it's own driver that can be shared by the MAC drivers.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 8152afff07d2768c6752a29ba4fd2a4bc76107b0)