iio: ad9680: register with the JESD204 framework
This is the start of the integration of this driver with the JESD204
framework.

If the jesd204_dev_register() returns a valid (non-NULL, non-err) pointer,
then don't continue with the normal setup.
This is broken for the current code; i.e. if the device registers with the
framework, things won't work, because some ops/function pointers are
required to be passed to the framework. This will be done in the next
commits.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>