jesd204: split link settings into active & staged
This change is an ABI change in itself, but it's an early stage anyway.
What it does, is to add support for having JESD204 link attributes
populated twice by having one group be 'active' and the next one 'staged'.

An example for `link0_active_num_lanes` & `link0_staged_num_lanes`:
- `link0_active_num_lanes` is always read-only, and reads the current
  number of lanes for the JESD204 link
- `link0_staged_num_lanes` can be read/written, but will only become active
  once it is committed; the mechanism is similar to `git commit` (in
  principle) and allows for multiple link-settings to be changed, and then
  committed/tested/validated/etc.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>