arch: arm: zc706: daq2/3: define vco max frequencies for CPLL
The ZC706 runs the DAQ2 & DAQ3 boards at a VCO frequency that is out of
range/spec. Configure a setting in the DT, to not allow an even higher
frequency that the one that was tested.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 19e54171c76bca26b3220ce3a9cae59e03a735b7)