arch: arm: e310/usrpe31x: add project tag + license
This project was created to showcase/demo the portability of the ADI Pluto
platform to similar projects based on AD936x.
e310 is one of them. This change updates the header to match it with the
HDL project.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit d4375660137c77173e9035d376cddb05c3c3dc62)