iio: hmc425a: rename hmc425 -> hmc425a
Mostly to prepare it for upstreaming, and avoid any questions of 'why name
some information hmc425 & not hmc425a?'.

Thing is, some drivers and parts get some more chips to support that have a
letter difference, and are really different, so it's good to be nitpick-y
about this detail.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 526bfe31fc5fb4ebafed3dfccbefba2901b86008)