ci,build: exit early if branch has an open PR
The situation is a bit complicated.
Travis-CI allows us to enable/disable builds for PRs & and for pushed
branches/tags.

If we disable builds for PRs, we won't get any build-runs from external
forks (external contributions).

If we disable builds for branches, people won't get a chance to have the CI
do a check before opening up a PR.

If we enable both, each re-spin of a PR requires up to 2 hours to do both
builds.

A compromise solution, is to implement a check on our build-script to exit
early if we detect that for a branch/tag we have an open PR. It's not
perfect, but should reduce some time.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 6af481dd0f3c4057ffd56a935f51e6fdc8cc29d0)