jesd204: move state transitions on a per-JESD204 link
This starts to add support for multi-link state transitions individual per
device, individual between each other.

Previously, state transitions were handled at the JESD204 top-level device.
This is bad for the case of a single device with multiple JESD204 links.

So, what this change does is to re-organize all state-transitions to be
kept on each JESD204 link.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>